docker-file
A Dockerfile is a text file that contains a set of instructions for building a custom Docker image. It defines the base image, environment settings, files to copy, commands to run, and services to start. Docker uses the Dockerfile to automatically create images with all the required dependencies and configurations for your application.